Better support for converting S3 Select JSON results back to parquet schemas
Timestamp & Date are loaded into pyarrow.lib.Table as strings. Then we attempt
to cast the table back to the original pyarrow schema of the data.  Currently,
this will likely fail all the time, but that is due to limited support in
pyarrow to cast strings back into pa.timestamp() and/or pa.date32()