Decoupled optional_where_clause behavior & bumped version
Previously, giving optional where clauses to lakeshack got passed down
to the metastore without checking to see if the metastore was created with
those columns listed as optional.

This meant that you couldn't filter results in lakehouse unless that column was
also present in the metastore. But there are times when you want to filter results
even if that column is not clustered well and won't reduce the number of files to
check. You shouldn't have to add that column to the metastore if it doesn't restrict
which files need to be searched.

bumped version to 0.2.2